feat(TUP-21457): License fetch feature with the https proxy https://jira.talendforge.org/browse/TUP-21457
